refactor(trn): split TrnService god-class to clear detekt TooManyFunctions
Extract TrnFinder (portfolio reads) + TrnPostProcessor (hydrate/migrate/secure)
and move settle/unsettle lifecycle into TrnSettlementService. TrnService now
holds writes only (26 -> 10 functions).
